bug 1162519 - use winrm for mach clobber on Windows. r=gps
UltraBlame original commit: a2c6a5972acba96ba70549261d44c102982c49e7

+++ python/mozbuild/mozbuild/base.py
@@ -295,12 +295,22 @@ class MozbuildObject(ProcessExecutionMixin):
             return False
         return Clobberer(self.topsrcdir, self.topobjdir).clobber_needed()
 
+    def have_winrm(self):
+        
+        p = subprocess.Popen(['winrm.exe', '-h'],
+                             stdout=subprocess.PIPE,
+                             stderr=subprocess.STDOUT)
+        return p.wait() == 1 and p.stdout.read().startswith('winrm')
+
     def remove_objdir(self):
         """Remove the entire object directory."""
 
-        
-        
-        rmtree(self.topobjdir.encode('utf-8'))
+        if sys.platform.startswith('win') and self.have_winrm():
+            subprocess.check_call(['winrm', '-rf', self.topobjdir])
+        else:
+            
+            
+            rmtree(self.topobjdir.encode('utf-8'))
 
     def get_binary_path(self, what='app', validate_exists=True, where='default'):
         """Obtain the path to a compiled binary for this build configuration.
